requirements-cpu.txt
# Requirements for CPU inference experiments
# Used by: code/laptop_benchmark.py, hardware_extended_platforms/scripts/
# Tested on: Intel i7-1165G7 (Windows 11), Intel Core Ultra 5/9, Raspberry Pi 5
#
# Install:
#   pip install -r requirements-cpu.txt

llama-cpp-python>=0.2.90
codecarbon>=2.3.1
pandas>=2.0.0
huggingface-hub>=0.23.0
